CHANGE: Per kilometre running costs are now deducted every kilometre instead of scaled running costs being deducted every tile. FIX: Odometer did not take account of reduced distance on diagonals. FIX: Stations with individually specified prices cost a negative amount to build (i.e., credited rather than debited the price from the player's bank balance) FIX: Occasional crash associated with route-finding for private cars. FIX: Overcrowding did not function.

+++ simconvoi.h
@@ -301,7 +301,7 @@ private:
 	// the odometer was last incremented.
 	// Used for converting tiles to km.
 	// @author: jamespetts
-	uint8 tiles_since_last_odometer_increment;
+	float tiles_since_last_odometer_increment;
 
 	/**
 	* Set, when there was a income calculation (avoids some cheats)
